ChatWindow implemented
I created the ChatWindow in tkinter to run in a separate thread. It's pretty slow so that gives Paul something to say in his part of the presentation.
I made some modifications to existing code to test that the chat works, the playerRoll function can be removed later but I'm just using to ensure
the chat can work as intended.

For any other tk windows (pre game windows) we won't need to thread since we won't be running them in parallel to pygame