Added __main__ to the Server folder to allow for -m running.
In fact, I have set it up so that using the -m flag to start the server is the only way it will allow.
Also in this commit have added the use of both stdout and stderr to output different messages.
Will update the docs on both the server and index to provide instructions on how to run the Server